feat: move `/generated` inside the `lib` folder
