+++ .gitignore
@@ -41,4 +41,4 @@ yarn-error.log*
 next-env.d.ts
 
 # Prisma
-/generated/
+/lib/generated/
