+++ eslint.config.mjs
@@ -12,7 +12,7 @@ const eslintConfig = [
       "@next/next/no-page-custom-font": "off",
       "@next/next/no-img-element": "off",
     },
-    ignorePatterns: ["generated/**/*"],
+    ignorePatterns: ["lib/generated/**/*"],
   }),
 ];
