+++ lib/actions/projects.ts
@@ -2,7 +2,7 @@
 
 import { z } from "zod";
 import { prisma } from "@/lib/prisma";
-import { Prisma, WorkspaceRole } from "@/generated/prisma/client";
+import { Prisma, WorkspaceRole } from "@/lib/generated/prisma/client";
 import { revalidatePath } from "next/cache";
 import { auth } from "@/lib/auth";
