+++ lib/actions/workspaces.ts
@@ -4,7 +4,7 @@ import { z } from "zod";
 import { prisma } from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 import { auth } from "@/lib/auth";
-import { WorkspaceRole } from "@/generated/prisma/client";
+import { WorkspaceRole } from "@/lib/generated/prisma/client";
 
 // Schema for workspace (workspace) creation/update validation
 const WorkspaceSchema = z.object({
