+++ prisma/schema.prisma
@@ -1,6 +1,6 @@
 generator client {
   provider = "prisma-client"
-  output = "../generated/prisma"
+  output = "../lib/generated/prisma"
 }
 
 datasource db {
